Dumped in Google seq2seq repo (github.com/google/seq2seq)
Massaged Cornell films data into two files (utterances.txt and
responses.txt) and setup a model config (as specified in the readme for
the repo at google.github.io/seq2seq). At the moment, should be able to
run this using their train.py script with the following parameters:

	python -m bin/train.py --config_paths="./example_configs/nmt_small.yml,
	./example_configs/train_seq2seq.yml,
	./example_configs/text_metrics_bpe.yml" --batch_size 32
	--train_steps 1000 --output_dir trained_models

Make sure you are in the working directory of $PROJECT_ROOT$/src/seq2seq
when invoking this. Model checkpoints will be dumped into
$PROJECT_ROOT$/src/seq2seq/trained_models, and pointing tensorboard at
this directory will allow you to monitor training. Some monitoring flags
were disabled by me at the moment due to a compatibility issue.

This is all messy as hell right now because it's mixed in with a bunch
of our other stuff, but this is just a proof of concept that this
library will do what we need.