Modified conv_context_encoder.py so that Context Encoder to load frozen subgraph
Utilize freeze_graph.py from tf utils to take a checkpoint and graphdef
protobuf (different from metagraph) from the cnn encoder, and produce a
frozen graphdef of the cnn with only nodes necesarry for inference and all
variables replaced with constant weights from checkpoint. Unfortunately,
for this to work I had to solve a heinous bug in the logic for
freeze_graph to ensure that nodes pertaining to lookup table initializers
(in this case for vocab) were kept. Once that was done, to ensure that the
lookup tables are initialized, they are added to the global collection
TABLE_INITIALIZERS so that the MonitoredSession initializes them. Yassss